% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_dim.R
\name{char.dim}
\alias{char.dim}
\title{Character Dimensions of Table}
\usage{
char.dim(obj, style, cx = 1)
}
\arguments{
\item{obj}{Table Object}

\item{style}{style}

\item{cx}{charcter expansion}
}
\description{
Given the table object and style object, will return the maximum width of columns, the height of characters in the units specified by user
}
