% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_dim.R
\name{char.width}
\alias{char.width}
\title{Character Width}
\usage{
char.width(obj, frmt, cx = 1)
}
\arguments{
\item{obj}{Table Object}

\item{frmt}{format object}

\item{cx}{shinkage parameter}
}
\description{
Returns the max width of all the elements in a vector in units, handles data types.
}
