% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_dim.R
\name{char.width1}
\alias{char.width1}
\title{Character Width}
\usage{
char.width1(vctr, frmt, cx = 1)
}
\arguments{
\item{vctr}{vector}

\item{frmt}{format object}

\item{cx}{shrinkage parameter}
}
\description{
Returns the max width in units, given gpar settings
}
