% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_struct.R
\name{colnames.row}
\alias{colnames.row}
\title{Column Names Row}
\usage{
colnames.row(colnames.obj)
}
\arguments{
\item{colnames.obj}{column names object}
}
\description{
Adjust index (reference number of rows above table for colnames) to account for line breaks
}
