\name{colnames.struct}
\alias{colnames.struct}
\title{Column Names Structure}
\usage{
  colnames.struct(col.names, linebreak = TRUE)
}
\arguments{
  \item{col.names}{vector of strings (result of
  colnames(data.frame))}

  \item{linebreak}{boolean}
}
\description{
  Creates a structure for printing column names and their
  hierachies
}

