% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional.struct.R
\name{conditional.struct}
\alias{conditional.struct}
\title{Conditional Structure}
\usage{
conditional.struct(data, byvars)
}
\arguments{
\item{data}{data.frame}

\item{byvars}{variables in addition to the group and label variables listed in the formula}
}
\description{
For variables defined as conditional variables in the formula, fuction to drive creating a list of simple table structures
}
