% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consect.struct.R
\name{consect.struct}
\alias{consect.struct}
\title{Consecutive Structure}
\usage{
consect.struct(vct)
}
\arguments{
\item{vct}{vector}
}
\description{
Given a vector returns several objects describing how(if) consecutive elements appear.  
This is a helper function for tbl.struct and is important in defining the column hierarchy.
}
