% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dborder.R
\name{dborder}
\alias{dborder}
\title{Draw Border}
\usage{
dborder(cord1, cord2, frmt)
}
\arguments{
\item{cord1}{Vector (x,y) indicating position of top left point of rectangle}

\item{cord2}{Vector (x,y) indicating position of bottom right point of rectangle}

\item{frmt}{Format}
}
\description{
Draws a border style arround a section of table
}
