\name{dprint-package}
\alias{dprint-package}
\alias{dprint}
\docType{package}
\title{Print Tabular Data to Graphics Device
}
\description{This package provides a generalized method for printing tabular data within the R environment in order to make the process of presenting high quality tabular output seamless for the useR.
}
\details{
\tabular{ll}{
Package: \tab dprint\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2011-11-10\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{Carlin Brickner
Maintainer: Carlin Brickner <carlin.brickner.rpackage@gmail.com>
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
