\name{dprint}
\alias{dprint}
\title{Print Table to Graphics Device (dprint)}
\usage{
  dprint(...)
}
\arguments{
  \item{...}{see dprint.table}
}
\description{
  drpint method (to be implemented)
}
\author{
  Carlin Brickner
}
\seealso{
  For now see \code{\link{dprint.table}}
}

