\name{dprint.simp}
\alias{dprint.simp}
\title{Print Table to Graphics Device (dprint)}
\usage{
  dprint.simp(tbl.obj, init = c(0, 6.9), style,
    char.dim.obj, size.simp.obj)
}
\arguments{
  \item{tbl.obj}{Simple Table object}

  \item{init}{Starting Position of Table}

  \item{style}{}

  \item{char.dim.obj}{}

  \item{size.simp.obj}{}
}
\description{
  drpint "simple table" class (to be implemented)
}

