% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dprint_simp.R
\name{dprint_simp}
\alias{dprint_simp}
\title{Print Table to Graphics Device (dprint)}
\usage{
dprint_simp(tbl.obj, init = c(0, 6.9), style, char.dim.obj, size.simp.obj)
}
\arguments{
\item{tbl.obj}{Simple Table object}

\item{init}{Starting Position of Table}

\item{style}{style object}

\item{char.dim.obj}{character dimension object}

\item{size.simp.obj}{size of text object passed to grid}
}
\description{
drpint "simple table" class (to be implemented)
}
