% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{Fb}
\alias{Fb}
\title{Format Binomial}
\usage{
Fb(prop = NULL, r = NULL, n = NULL, digits.p = 4, nsmall.p = 2,
  digits.n = 6)
}
\arguments{
\item{prop}{column name representing a proportion}

\item{r}{column name representing number of binomial events observed}

\item{n}{column name representing number of observations}

\item{digits.p}{digits for proportion presentation. see format}

\item{nsmall.p}{nsmall for proportion presentation.}

\item{digits.n}{number of digits for n presentation}
}
\description{
Inline formula formatting of columns representing binomial summary statistics
}
