% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{Fc}
\alias{Fc}
\title{Format Continuous}
\usage{
Fc(mn, std, digits = 2, nsmall = 2, NAmiss = TRUE)
}
\arguments{
\item{mn}{column name representing a mean}

\item{std}{column name representing standard deviation (Variance, Margin of Error, etc.)}

\item{digits}{see format}

\item{nsmall}{see format}

\item{NAmiss}{present NA as white space}
}
\description{
Inline formula formatting of columns representing summary statistics for Continuous data
}
