% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{Fci}
\alias{Fci}
\title{Format Interval}
\usage{
Fci(cil, ciu, mn = NA, digits = 2, nsmall = 2, NAmiss = TRUE)
}
\arguments{
\item{cil}{column name representing lower interval}

\item{ciu}{column name representing upper interval}

\item{mn}{column name representing a mean}

\item{digits}{see format}

\item{nsmall}{see format}

\item{NAmiss}{present NA as white space}
}
\description{
Inline formula formatting of columns representing lower and upper bound of an interval where the results are of the form 'mn (cil, ciu)'
}
