% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitpage.R
\name{fitpage}
\alias{fitpage}
\title{Fit Table to Page}
\usage{
fitpage(size.simp, pagelayout, fit.width = FALSE, fit.height = FALSE,
  fit = FALSE, cex.0)
}
\arguments{
\item{size.simp}{size object}

\item{pagelayout}{dimensions of page}

\item{fit.width}{boolean, force to fit to width}

\item{fit.height}{boolean, force to fit height}

\item{fit}{boolean, force to fit minimum of height and wdith}

\item{cex.0}{default cex}
}
\description{
Calcuates character expansion to apply accross table presentation so that table fits within margins
}
