\name{fmla.inter}
\alias{fmla.inter}
\title{Formula Interface Describing the Structure of Tabular Data}
\usage{
  fmla.inter(f, data = NULL, regx = NA, ...)
}
\arguments{
  \item{f}{formula}

  \item{data}{data.frmae}

  \item{regx}{remove regular expression from column name}

  \item{...}{}
}
\description{
  Parse the left and right hand side of the formula.
}

