\name{fmla.rhs}
\alias{fmla.rhs}
\title{Formula Interface: Right Hand Side}
\usage{
  fmla.rhs(f, span,
    Rnl = list(Rn.o = list(), Rn.n = NULL, rn.i = 0),
    byvars = list(byvars1 = NULL, byvars2 = NULL, byvars.i = 0))
}
\arguments{
  \item{f}{formula}

  \item{span}{There are two types of calls, when TRUE
  returns the spanning text, does not when FALSE}

  \item{Rnl}{}

  \item{byvars}{}
}
\description{
  Parses Right hand side of formula
}

