% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmla_inter.R
\name{fmla_lhs}
\alias{fmla_lhs}
\title{Formula Interface: Left Hand Side}
\usage{
fmla_lhs(f, grp = NULL, lvl = NULL, rhs.lpad = FALSE)
}
\arguments{
\item{f}{formula}

\item{grp}{character, name of variable which groups rows together}

\item{lvl}{character, name of variable which labels each row}

\item{rhs.lpad}{boolean, right pad}
}
\description{
Parses Left hand side of formula
}
\author{
Rocco Napoli
}
