% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmla_inter.R
\name{fmla_rhs}
\alias{fmla_rhs}
\title{Formula Interface: Right Hand Side}
\usage{
fmla_rhs(f, span, Rnl = list(Rn.o = list(), Rn.n = NULL, rn.i = 0),
  byvars = list(byvars1 = NULL, byvars2 = NULL, byvars.i = 0))
}
\arguments{
\item{f}{formula}

\item{span}{There are two types of calls, when TRUE returns the spanning text, does not when FALSE}

\item{Rnl}{Rename}

\item{byvars}{Condition on these variables}
}
\description{
Parses Right hand side of formula
}
\author{
Rocco Napoli
}
