% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_sheet.R
\name{frmt}
\alias{frmt}
\title{Format of Table}
\usage{
frmt(fontfamily = NULL, fontface = NULL, fontsize = NULL,
  text.all = FALSE, col = NULL, bg = NULL, buf = NULL, bty = NULL,
  lwd = NULL, lty = NULL, lcol = NULL, linespace = NULL)
}
\arguments{
\item{fontfamily}{see gpar}

\item{fontface}{see gpar}

\item{fontsize}{see gpar}

\item{text.all}{If TRUE, when frmt.bdy is defined and nothing else, fontfamily, fontface, fontsize, col will be applied to frmt.col and frmt.row to avoid excessive parameters. Need to define some general rules - probably not implemented correct at this time.}

\item{col}{Color of text}

\item{bg}{Background Color}

\item{buf}{number of character space (buffer) to be forced inbetween columns, and after the labels and group display.  When left and right justified, .5 buffer is added and subtracted to respective side of borders}

\item{bty}{style for box around section of table. "=" = above and below, "o" = completely enclosed, "_"=underneath}

\item{lwd}{line size for bty, -1 suppresses}

\item{lty}{line type for bty}

\item{lcol}{color of lines}

\item{linespace}{Number of lines between columns}
}
\description{
Creates a data structure to define the gpar settings passed to the various grid.* function calls, 
as well as other formatting controls
}
