\name{ftr}
\alias{ftr}
\title{Footer}
\usage{
  ftr(txt1,
    frmt1 = frmt(fontfamily = "", fontface = "plain", fontsize = 8, col = "black", linespace = 0.75),
    page = TRUE, date = TRUE,
    pagelayout.obj = pagelayout(dtype = "portrait", margins = c(1, 0.5)),
    pgtxt2 = "page", pagenum = NULL)
}
\arguments{
  \item{txt1}{}

  \item{frmt1}{}

  \item{page}{}

  \item{date}{}

  \item{pagelayout.obj}{}

  \item{pgtxt2}{}

  \item{pagenum}{}
}
\description{
  Footer
}

