% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdr.R
\name{hdr}
\alias{hdr}
\title{Header}
\usage{
hdr(txt1, frmt1 = frmt(fontfamily = "", fontface = "bold", fontsize = 20, col
  = "blue", bty = "_", lwd = 2, linespace = 1), txt2 = NA,
  frmt2 = frmt(fontfamily = "", fontface = "bold", fontsize = 16, col = "red",
  bty = "_", lwd = 2, linespace = 1), pagelayout.obj = pagelayout(dtype =
  "portrait", margins = c(1, 0.5)))
}
\arguments{
\item{txt1}{Text 1}

\item{frmt1}{Format 1}

\item{txt2}{Text 2}

\item{frmt2}{Format 2}

\item{pagelayout.obj}{Layout Object}
}
\description{
Header
}
