% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.grp.R
\name{insert.grp}
\alias{insert.grp}
\title{Insert Values Table Structure}
\usage{
insert.grp(tbl.obj)
}
\arguments{
\item{tbl.obj}{Table Object}
}
\description{
Dispatcher for insert.grp1. Inserts NA positions for tbl, label, and group structure from tbl.obj.
}
