% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kill.multiregx.R
\name{kill.multiregx}
\alias{kill.multiregx}
\title{Kill Multiple Regular Expression}
\usage{
kill.multiregx(string, regx)
}
\arguments{
\item{string}{Input string which will have regular expressions removed}

\item{regx}{Perl Regular expression}
}
\description{
Compacting some regular expression logic to remove regular expressions from vector
}
