% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagelayout.R
\name{pagelayout}
\alias{pagelayout}
\title{Page Layout}
\usage{
pagelayout(dtype = "rgraphics", margins = NULL, pg.dim = NULL)
}
\arguments{
\item{dtype}{Device type, currently available "rgraphics", "portrait", "landscape"}

\item{margins}{A numerical vector of the form c(bottom, left, top, right) which gives the margin size specified in inches.}

\item{pg.dim}{If custom page dimensions define a vector of c(width, height) units}
}
\description{
Page Layout
}
