% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_descriptives.R
\name{rdesc}
\alias{rdesc}
\title{Simulate Descriptives: Continuous}
\usage{
rdesc(n.grp = 8, n.lvls = 5, n.grp2 = NULL, rnd = FALSE)
}
\arguments{
\item{n.grp}{number of row groups}

\item{n.lvls}{number of rows in each row group}

\item{n.grp2}{another row grouping, such that grp is subset of grp2}

\item{rnd}{boolean round results, data frame then returns as character}
}
\description{
Simulate a Data Frame containing descriptive statistics summarizing a continuous data
}
