\name{row.hl}
\alias{row.hl}
\title{Row Highlight}
\usage{
  row.hl(dx, col = "yellow")
}
\arguments{
  \item{dx}{index rows of data.frame to highlight}

  \item{col}{color}
}
\description{
  Creates a data structure for highlighting rows of the
  table based on indices provided by the user
}

