% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablebreak.R
\name{tablebreak}
\alias{tablebreak}
\title{Table Break}
\usage{
tablebreak(tbl.obj.l, cur.tbl, dx)
}
\arguments{
\item{tbl.obj.l}{list of table objects}

\item{cur.tbl}{Index to current table}

\item{dx}{Row to break table on}
}
\description{
Given an index, breaks table into multiple tables
}
