% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.struct.R
\name{tbl.struct}
\alias{tbl.struct}
\title{Table Structure}
\usage{
tbl.struct(fmla = NULL, data, label = NULL, group = NULL, regx = NA,
  main = NA, footnote = NA, row.hl = list(dx = NULL, col = NULL))
}
\arguments{
\item{fmla}{Formula interface to define table structure}

\item{data}{data.frame}

\item{label}{name of column containing row labels}

\item{group}{name of column containing hieriarchy labels for the row names}

\item{regx}{regular expression to be removed from original column names}

\item{main}{Table title}

\item{footnote}{footnote}

\item{row.hl}{row highlight object see row.hl function}
}
\description{
Generalization of table structure
}
