% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.struct.simp.R
\name{tbl.struct.simp}
\alias{tbl.struct.simp}
\title{Simple Table Structure}
\usage{
tbl.struct.simp(data, label = NULL, group = NULL, main = NA,
  footnote = NA, colnames.obj = NULL)
}
\arguments{
\item{data}{data.frame}

\item{label}{name of column containing row labels}

\item{group}{name of column containing hieriarchy labels for the row names}

\item{main}{table title}

\item{footnote}{foot note}

\item{colnames.obj}{colnames object defined by colnames.struct}
}
\description{
Simple Table Structure
}
