\name{vector.linebreak}
\alias{vector.linebreak}
\title{Vector Line Break}
\usage{
  vector.linebreak(vctr)
}
\arguments{
  \item{vctr}{character vector}
}
\description{
  To break vectors apart around an escape character
  [backslash]n that indicates a line break. This is
  designed for handling line breaks for the table title
  (main) and the footnotes
}

