% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.struct.R
\name{vector.struct}
\alias{vector.struct}
\title{Vector Structure}
\usage{
vector.struct(vctr = NA)
}
\arguments{
\item{vctr}{character vector}
}
\description{
Defines the structure of a vector given to footnote, main, etc.
}
