% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_dim.R
\name{char.height}
\alias{char.height}
\title{Character Height}
\usage{
char.height(charact = "A", frmt, cx = 1)
}
\arguments{
\item{charact}{Table Object}

\item{frmt}{format object}

\item{cx}{shinkage parameter}
}
\description{
Returns height in units(inches) based on a gpar setting
}
