% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames_struct.R
\name{colnames.linebreak}
\alias{colnames.linebreak}
\title{Column Names Line Break}
\usage{
colnames.linebreak(colnames.obj)
}
\arguments{
\item{colnames.obj}{column names object}
}
\description{
Inserts a line break in column names when the escape character [backslash] n is found.
}
