% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dprint.R
\name{dprint}
\alias{dprint}
\title{Print Table to Graphics Device (dprint)}
\usage{
dprint(...)
}
\arguments{
\item{...}{see dprint.table}
}
\description{
drpint method (to be implemented)
}
\seealso{
For now  see \code{\link{dprint.table}}
}
\author{
Carlin Brickner
}
