% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dprint.R
\name{dprint}
\alias{dprint}
\title{Print Table to Graphics Device (dprint)}
\usage{
dprint(data, fmla = NULL, label = NULL, group = NULL, regx = NA,
  style = NULL, main = NA, footnote = NA, dtype = "rgraphics",
  pg.dim = NULL, margins = NULL, showmargins = FALSE, row.hl = NULL,
  fit.width = FALSE, fit.height = FALSE, fit = FALSE, newpage = FALSE,
  center.horz = FALSE, center.vert = FALSE, center = FALSE,
  f.hdr = NULL, f.ftr = NULL, pagenum = NULL, lastcall = NULL)
}
\arguments{
\item{data}{input data.frame of class "data.frame"}

\item{fmla}{an object of class "formula". Formula interface to define table structure. See fmla function}

\item{label}{instead of using fmla, the name of column containing row labels}

\item{group}{instead of using fmla,name of column containing hieriarchy labels for the row names}

\item{regx}{regular expression used to remove unwanted text displayed from original column names (e.g. merge applied with .x and .y appended to duplicate column nammes)}

\item{style}{style sheet object.  See style function}

\item{main}{table title}

\item{footnote}{footnote}

\item{dtype}{Graphics device type referred to by names, sets default page settings.Device type, currently available "rdevice", "portrait", "landscape"}

\item{pg.dim}{a vector of c(height,width) units used to describe the dimensinos of a custom page and  over ride dtype.  When printing to a multiple page pdf with custom dimensions, dtype should be set to some character other than "redevice" (i.e. "custom") because dev.new() will be used to start a new window instead of grid.text()}

\item{margins}{A numerical vector of the form c(bottom, left, top, right) which gives the margin size specified in inches. Other forms include a constant for all margins and c(top/bottom,left/right)}

\item{showmargins}{Display margins on page in red. Usefull for tinkering with presentation}

\item{row.hl}{row hightlight object. Conditional highlight; row.hl(dx=, col=)}

\item{fit.width}{logical. If TRUE, Forces the table to fit the table horizontally}

\item{fit.height}{logical. If TRUE, Forces the table to fit the table vertically, if conditional variabled defined will force the first table to fit one page and same ratio will be applied to all tables}

\item{fit}{logical. If TRUE, Force the table to fit both horizontally&vertically (smaller of two cex calulcations)}

\item{newpage}{logical. If TRUE, when the device runs out of space a new page will automatically be started. DEsigned for multipage pdf reports.}

\item{center.horz}{boolean, Center table horizontally}

\item{center.vert}{boolean, Center table vertically, should only be used on one table}

\item{center}{boolean, Center both vertically and horizontally. These should probably not be used with the fit.* parameters}

\item{f.hdr}{pass Function for printing header}

\item{f.ftr}{pass Function for printing Footer}

\item{pagenum}{starting page number, will override pagenumber from lastcall}

\item{lastcall}{object returned from last call from dprint.  Can use this as reference for a second table presented on the same device as a pervious.  dprint continues printing to device with fixed separation between tables}
}
\description{
Print table structure, currently provided as an input data frame, to the Graphics Device. Specification of the layout of the table is done through a formula interface. In addition, formula interface allows user to apply inline custom formats, renaming, and any mathematical expressions supported by model.matrix
}
\author{
Carlin Brickner
}
