% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dprint.R
\name{dprint}
\alias{dprint}
\title{Print Table to Graphics Device (dprint)}
\usage{
dprint(data, ...)
}
\arguments{
\item{data}{Input data.frame of class "data.frame"}

\item{...}{see \code{\link{dprint.data.frame}}}
}
\description{
drpint method (to be implemented)
}
\author{
Carlin Brickner
}
