\name{fitpage}
\alias{fitpage}
\title{Fit Table to Page}
\usage{
  fitpage(size.simp, pagelayout, fit.width = FALSE,
    fit.height = FALSE, fit = FALSE, cex.0)
}
\arguments{
  \item{size.simp}{}

  \item{pagelayout}{}

  \item{fit.width}{}

  \item{fit.height}{}

  \item{fit}{}

  \item{cex.0}{default cex}
}
\description{
  Calcuates character expansion to apply accross table
  presentation so that table fits within margins
}

