% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmla_inter.R
\name{fmla_inter}
\alias{fmla_inter}
\title{Formula Interface Describing the Structure of Tabular Data}
\usage{
fmla_inter(f, data = NULL, regx = NA)
}
\arguments{
\item{f}{formula}

\item{data}{data.frmae}

\item{regx}{remove regular expression from column name}
}
\description{
Parse the left and right hand side of the formula.
}
\author{
Rocco Napoli
}
