\name{fmla_lhs}
\alias{fmla_lhs}
\title{Formula Interface: Left Hand Side}
\usage{
  fmla_lhs(f, grp = NULL, lvl = NULL, rhs.lpad = FALSE)
}
\arguments{
  \item{f}{formula}

  \item{grp}{}

  \item{lvl}{}

  \item{rhs.lpad}{}
}
\description{
  Parses Left hand side of formula
}
\author{
  Rocco Napoli
}

