% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{Fr}
\alias{Fr}
\title{Format Round}
\usage{
Fr(x, digits = 2, nsmall = 2, NAmiss = TRUE)
}
\arguments{
\item{x}{column name}

\item{digits}{see format}

\item{nsmall}{see format}

\item{NAmiss}{present NA as white space}
}
\description{
Abreviation of rouding function
}
