\name{hdr}
\alias{hdr}
\title{Header}
\usage{
  hdr(txt1,
    frmt1 = frmt(fontfamily = "", fontface = "bold", fontsize = 20, col = "blue", bty = "_", lwd = 2, linespace = 1),
    txt2 = NA,
    frmt2 = frmt(fontfamily = "", fontface = "bold", fontsize = 16, col = "red", bty = "_", lwd = 2, linespace = 1),
    pagelayout.obj = pagelayout(dtype = "portrait", margins = c(1, 0.5)))
}
\arguments{
  \item{txt1}{}

  \item{frmt1}{}

  \item{txt2}{}

  \item{frmt2}{}

  \item{pagelayout.obj}{}
}
\description{
  Header
}

