\name{kill.multiregx}
\alias{kill.multiregx}
\title{Kill Multiple Regular Expression}
\usage{
  kill.multiregx(string, regx)
}
\arguments{
  \item{string}{}

  \item{regx}{Perl Regular expression}
}
\description{
  Compacting some regular expression logic to remove
  regular expressions from vector
}

