\name{size.simp}
\alias{size.simp}
\title{Size of Simple Table}
\usage{
  size.simp(tbl.obj, char.dim.obj, pagelayout, loc.y)
}
\arguments{
  \item{tbl.obj}{}

  \item{char.dim.obj}{}

  \item{pagelayout}{}

  \item{loc.y}{}
}
\description{
  Measures size of properties of the simple table
}

